LayoutTests/http/tests/media/media-source/video-media-source-state-changes.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script src="/media/resources/media-source/webm/segment-info.js"></script>
        <script src="media-source.js"></script>
        <script>
            var segmentHelper = new MediaSourceTest.SegmentHelper(WebMSegmentInfo.testWebM);
            var seekTime;

            function appendMediaSegmentsForSeek(videoTag, segmentSeekTime)
            {
                var index = segmentHelper.getMediaSegmentIndexForTimestamp(segmentSeekTime);
                var endIndex = index + 2;
                var totalMediaSegments = segmentHelper.mediaSegments.length;

                for (; (index < endIndex) && (index < totalMediaSegments); index++) {
                    segmentHelper.appendMediaSegment(index);
                }

                if (index >= totalMediaSegments) {
                    videoTag.webkitSourceEndOfStream(HTMLMediaElement.EOS_NO_ERROR);
                    return;
                }
            }

            function onSourceEvent(event)
            {
                consoleWrite('EVENT(' + event.type + ') : ' + MediaSourceTest.getSourceStateName(event.target.webkitSourceState));
            }

            function onFirstSourceOpen(event)
            {
                consoleWrite("onFirstSourceOpen");

                // This is the first time the source is opened. We just want to append
                // the initialization segment and all media segments until the end of stream.
                // This is testing the
                // SOURCE_CLOSED -> SOURCE_OPEN -> SOURCE_ENDED transition path.

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceopen', onFirstSourceOpen);
                segmentHelper.addSourceId();

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                segmentHelper.appendInitSegment();

                videoTag.addEventListener('webkitsourceended', onFirstSourceEnded);

                // Start appending media segments.
                segmentHelper.appendUntilEndOfStream(0);
            }

            function onFirstSourceEnded(event)
            {
                consoleWrite("onFirstSourceEnded");
                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceended', onFirstSourceEnded);

                MediaSourceTest.expectSourceState(videoTag, videoTag.SOURCE_ENDED);

                videoTag.addEventListener('playing', triggerFirstSeek);
            }

            function triggerFirstSeek(event)
            {
                consoleWrite("triggerFirstSeek");

                var videoTag = event.target;
                videoTag.removeEventListener('playing', triggerFirstSeek);

                // Now that we have appended all of the media segments and ended the stream,
                // lets do a seek to test the SOURCE_ENDED -> SOURCE_OPEN transition.
                videoTag.addEventListener('seeking', onFirstSeeking);
                videoTag.addEventListener('seeked', onFirstSeeked);
                videoTag.currentTime = seekTime;
            }

            function onFirstSeeking(event)
            {
                consoleWrite("onFirstSeeking");

                var videoTag = event.target;
                videoTag.removeEventListener('seeking', onFirstSeeking);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                // Start appending media segments.
                appendMediaSegmentsForSeek(videoTag, videoTag.currentTime);
            }

            function onFirstSeeked(event)
            {
                consoleWrite("onFirstSeeked");

                var videoTag = event.target;
                videoTag.removeEventListener('seeked', onFirstSeeked);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                // Now it is time to test seeking while still in SOURCE_OPEN.
                videoTag.addEventListener('seeking', onSecondSeeking);
                videoTag.addEventListener('seeked', onSecondSeeked);
                videoTag.currentTime = seekTime;
            }

            function onSecondSeeking(event)
            {
                consoleWrite("onSecondSeeking");

                var videoTag = event.target;
                videoTag.removeEventListener('seeking', onSecondSeeking);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                // Start appending media segments.
                appendMediaSegmentsForSeek(videoTag, videoTag.currentTime);
            }

            function onSecondSeeked(event)
            {
                consoleWrite("onSecondSeeked");
                var videoTag = event.target;
                videoTag.removeEventListener('seeked', onSecondSeeked);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                // Test SOURCE_OPEN -> SOURCE_CLOSED -> SOURCE_OPEN transition path.
                videoTag.addEventListener('webkitsourceclose', onFirstSourceClose);
                MediaSourceTest.setSrcToMediaSourceTestURL(videoTag);
            }

            function onFirstSourceClose(event)
            {
                consoleWrite("onFirstSourceClose");

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceclose', onFirstSourceClose);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_CLOSED);

                videoTag.addEventListener('webkitsourceopen', onSecondSourceOpen);
            }

            function onSecondSourceOpen(event)
            {
                consoleWrite("onSecondSourceOpen");

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceopen', onSecondSourceOpen);
                segmentHelper.addSourceId();

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                segmentHelper.appendInitSegment();

                videoTag.addEventListener('webkitsourceended', onSecondSourceEnded);

                // Start appending media segments.
                segmentHelper.appendUntilEndOfStream(0);
            }

            function onSecondSourceEnded(event)
            {
                consoleWrite("onSecondSourceEnded");

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceended', onSecondSourceEnded);

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_ENDED);
                //MediaSourceTest.expectReadyState(videoTag, HTMLMediaElement.HAVE_NOTHING);

                videoTag.addEventListener('playing', triggerSecondSourceClose);
            }

            function triggerSecondSourceClose(event)
            {
                consoleWrite("triggerSecondSourceClose");

                var videoTag = event.target;
                videoTag.removeEventListener('playing', triggerSecondSourceClose);

                // Test SOURCE_ENDED -> SOURCE_CLOSED -> SOURCE_OPEN transition path.
                videoTag.addEventListener('webkitsourceclose', onSecondSourceClose);
                MediaSourceTest.setSrcToMediaSourceTestURL(videoTag);
            }

            function onSecondSourceClose(event)
            {
                consoleWrite("onSecondSourceClose");

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceclose', onSecondSourceClose);
                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_CLOSED);

                videoTag.addEventListener('webkitsourceopen', onThirdSourceOpen);
            }

            function onThirdSourceOpen(event)
            {
                consoleWrite("onThirdSourceOpen");

                var videoTag = event.target;
                videoTag.removeEventListener('webkitsourceopen', onThirdSourceOpen);
                segmentHelper.addSourceId();

                MediaSourceTest.expectSourceState(videoTag, HTMLMediaElement.SOURCE_OPEN);

                endTest();
            }

            function onLoad()
            {
                findMediaElement();

                waitForEventAndFail('error');
                waitForEvent('loadstart');
                waitForEvent('playing');
                waitForEvent('seeking');
                waitForEvent('seeked');
                waitForEvent('ended');
                waitForEvent('emptied');

                video.addEventListener('webkitsourceopen', onSourceEvent);
                video.addEventListener('webkitsourceended', onSourceEvent);
                video.addEventListener('webkitsourceclose', onSourceEvent);

                video.addEventListener('webkitsourceopen', onFirstSourceOpen);

                MediaSourceTest.logSourceState(video);
                segmentHelper.init(video, function(success)
                {
                    if (!success) {
                        failTest("Failed to load segment data");
                        return;
                    }

                    MediaSourceTest.setSrcToMediaSourceTestURL(video);
                });
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video autoplay> </video>
        <p>Tests MediaSource API state transitions.</p>
    </body>
</html>
